requirements.txt
python-telegram-bot==20.7
geopy==2.4.1
googlemaps==4.10.0
